feat:动画、深夜模式

+++ TodoList/Views/ListView.swift
@@ -14,25 +14,32 @@ struct ListView: View {
     @EnvironmentObject var viewModel: TodoListViewModel
     
     var body: some View {
-        List{
-            // 方法一：捕获 `$0` 放在闭包外
-//            ForEach(viewModel.todoItems){
-//                // 这里的 `$0` 属于 ForEach 的闭包参数（第一层闭包）
-//                let item = $0
-//                TodoListRow(item: item) // $0 代表当前遍历的 item
-//                    .onTapGesture {
-//                        viewModel.updateItem(item: item)
-//                    }
-//            }
-            // 方法二：明确命名闭包参数为 `item`
-            ForEach(viewModel.todoItems) { item in
-                TodoListRow(item: item)
-                    .onTapGesture {
-                        viewModel.updateItem(item: item)  // ✅ 直接使用已命名的参数
-                    }
+        ZStack{
+            if viewModel.todoItems.isEmpty {
+                NoItemsView()
             }
-            .onDelete(perform: viewModel.deleteItem)
-            .onMove(perform: viewModel.moveItem)
+            else{
+                List{
+                    // 方法一：捕获 `$0` 放在闭包外
+        //            ForEach(viewModel.todoItems){
+        //                // 这里的 `$0` 属于 ForEach 的闭包参数（第一层闭包）
+        //                let item = $0
+        //                TodoListRow(item: item) // $0 代表当前遍历的 item
+        //                    .onTapGesture {
+        //                        viewModel.updateItem(item: item)
+        //                    }
+        //            }
+                    // 方法二：明确命名闭包参数为 `item`
+                    ForEach(viewModel.todoItems) { item in
+                        TodoListRow(item: item)
+                            .onTapGesture {
+                                viewModel.updateItem(item: item)  // ✅ 直接使用已命名的参数
+                            }
+                    }
+                    .onDelete(perform: viewModel.deleteItem)
+                    .onMove(perform: viewModel.moveItem)
+                }
+            }
         }
         .listStyle(.plain)
         .navigationTitle("Todo List APP")
